Implement simple shell with typing
Most of this is just temp scaffolding. What we
really need is to implement process management / thread management /
scheduling next, so that we can run the shell as its own program, in
true microkernel fashion.

(Also, will need to implement IPC messaging, so that the user space
processes can communicate. The keyboard driver itself will also be a
process, so the keyboard handling/buffering/scan code decoding is also
something that will need to be redone)